Make colors match Super Famicom buttons, match font to rest of skin, m
Existing skin had button colors that don't match the standard Super Famicom color scheme for A, B, X, and Y (which is used across most retro consoles). In addition, the font on these buttons was changed to match the font used on the rest of the skin and make it easier to read.